requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
psycopg2
pymongo
python-dotenv
